+++ plugin.py
@@ -228,13 +228,47 @@ class PHPUnit():
         php_executable = view.settings().get('phpunit.php_executable')
         if php_executable:
             php_executable = os.path.expanduser(php_executable)
+
             if not os.path.isfile(php_executable):
                 return sublime.status_message('PHPUnit: PHP executable not found')
+
             if not os.access(php_executable, os.X_OK):
                 return sublime.status_message('PHPUnit: PHP executable is not executable')
+
             cmd += php_executable + ' '
 
             debug_message('PHP executable: %s' % php_executable)
+
+        else:
+            php_versions_path = view.settings().get('phpunit.php_versions_path')
+            if php_versions_path:
+                php_versions_path = os.path.expanduser(php_versions_path)
+                if not os.path.isdir(php_versions_path):
+                    return sublime.status_message('PHPUnit: PHP versions path does not exist or is an invalid directory')
+
+                php_version_file = os.path.join(working_dir, '.php-version')
+                if os.path.isfile(php_version_file):
+
+                    with open(php_version_file, 'r') as f:
+                        php_version = f.read().strip()
+
+                    php_version = re.match('^master|[1-9]+\.[0-9x](?:\.[0-9x])?(?:snapshot)?$', php_version)
+                    if not php_version:
+                        return sublime.status_message('PHPUnit: PHP version file version is malformed')
+
+                    php_version = php_version.group(0)
+
+                    php_executable = os.path.join(php_versions_path, php_version, 'bin', 'php')
+
+                    if not os.path.isfile(php_executable):
+                        return sublime.status_message('PHPUnit: PHP executable for version file not found')
+
+                    if not os.access(php_executable, os.X_OK):
+                        return sublime.status_message('PHPUnit: PHP executable for version file is not executable')
+
+                    cmd += php_executable + ' '
+
+                    debug_message('PHP executable: %s' % php_executable)
 
         if view.settings().get('phpunit.composer') and os.path.isfile(os.path.join(working_dir, os.path.join('vendor', 'bin', 'phpunit'))):
             executable = os.path.join(working_dir, os.path.join('vendor', 'bin', 'phpunit'))